refactor: retire aether-player-component, inline tile-step into sokoban (issue 198)
The player component was a half-baked abstraction — two motion modes,
a control-mail surface for position/velocity/mode, and a request/reply
round-trip with the world authority — that hadn't earned its keep.
Sokoban was the only consumer, and inlining the tile-step bits drops
both crates' worth of complexity for one direct Key handler plus a
render call.

- Inline tile-step into sokoban: Key handler, player triangle render,
  CameraTopdownSet follow per tick.
- Delete crates/aether-player-component/.
- Delete the five aether.player.* kinds (set_position, set_velocity,
  set_mode, request_step, step_result) — sokoban talks to itself now.
- Drop the workspace member.
- Smoke-tested live: D-press pushes the player right, the box lands
  on its target, camera follows.

Closes #198 — the z-axis ask was extending the player abstraction
that's now retired; revisit if/when a player concept comes back.
