+++ Cargo.toml
@@ -11,7 +11,6 @@ members = [
     "crates/aether-mesh",
     "crates/aether-camera-component",
     "crates/aether-mesh-editor-component",
-    "crates/aether-player-component",
     "crates/aether-static-mesh-component",
     "crates/aether-hub-protocol",
     "crates/aether-mail",
