--- crates/aether-player-component/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "aether-player-component"
-version.workspace = true
-edition.workspace = true
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-aether-component = { path = "../aether-component" }
-aether-kinds = { path = "../aether-kinds" }
